Backend_TRL/python_scripts/requirements.txt
# Procesamiento de documentos
pymupdf>=1.24.0
pytesseract>=0.3.10
pillow>=10.0.0
unstructured>=0.12.0
python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1.2
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Base de datos
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Utilidades
python-dotenv>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0

# Google AI
google-genai>=0.3.0
google-generativeai>=0.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
